bug fixing on all pages, transaction, ticket details, search, payment, home, event, discover
