+++ src/pages/event-page/eventCreate-page/components/schema.ts
@@ -1,30 +1,8 @@
 import * as Yup from "yup";
-
-// Define the form values interface
-export interface EventFormValues {
-  name: string;
-  start_date: string;
-  start_time: string;
-  end_date: string;
-  end_time: string;
-  description: string;
-  location: string;
-  price: number;
-  total_seats: number;
-  category: string;
-  // Add voucher fields
-  create_voucher: boolean;
-  voucher_code: string;
-  discount_amount: number;
-  voucher_start_date: string;
-  voucher_start_time: string;
-  voucher_end_date: string;
-  voucher_end_time: string;
-  max_usage: number;
-}
+import { IEventFormValues } from "./types";
 
 // Initial form values with correct case for category
-export const eventInitialValues: EventFormValues = {
+export const eventInitialValues: IEventFormValues = {
   name: "",
   start_date: "",
   start_time: "",
