+++ src/pages/home-page/sections/events-section/components/types.tsx
@@ -1,0 +1,9 @@
+export interface IEventCard {
+  id: string;
+  name: string;
+  event_image: string;
+  location: string;
+  start_date: string;
+  end_date: string;
+  description?: string;
+}
