+++ src/pages/home-page/sections/events-section/concert/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import EventCard from "@/components/cards/eventCard";
-import { IEventCard } from "../components";
+import { IEventCard } from "../components/types";
 import axios from "axios";
 import Link from "next/link";
 import { cloudinaryBaseUrl } from "@/components/config/cloudinary";
