+++ src/pages/payment-page/paymentSuccess-page/components/types.ts
