+++ src/pages/search-page/components/type.ts
