+++ src/pages/ticket-details-page/components/types.ts
